+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/events/Events.kt
@@ -1,10 +1,6 @@
 package slatekit.jobs.events
 
 import slatekit.common.Status
-import slatekit.functions.policy.Policy
-import slatekit.jobs.WorkRequest
-import slatekit.jobs.WorkResult
-import slatekit.jobs.Worker
 
 
 interface Events<T> {
@@ -25,11 +21,3 @@ interface Events<T> {
 
 
 
-interface WorkEvents : Events<Worker<*>> {
-
-    /**
-     * Applies the policy to these workers
-     * Policies add behaviour to workers such as retries, limits, error ratios, etc
-     */
-    fun apply(policy: Policy<WorkRequest, WorkResult>)
-}
